fix: unify edting layer component
